Populate image preview toolbar content early
This reduces the flicker that we see when selecting an image on the
Network panel for preview. The reason for the flicker is the
asynchronicity between creating the toolbar and populating its content.

There is still a slight flicker left over that is caused by removing the
widget and then asynchronously later creating a new widget when changing
the selection. A lot more re-architecturing would be required to fix
this.

R=aerotwist@chromium.org

Fixed: chromium:1142729
Change-Id: Idefd4731577d6eada8ab0a15e7be2d5ee924e46e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2510393
Commit-Queue: Yang Guo <yangguo@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Auto-Submit: Yang Guo <yangguo@chromium.org>