Add dark mode patch option to `createShadowRootWithCoreStyles`
This CL follows on from https://crrev.com/c/2516345 and adds the
`enableLegacyPatching` flag to `createShadowRootWithCoreStyles`. This
was harder as the function already took optional argumetns so I have
updated it to take an options bag, and then added an additional option.

This CL means that all the sites where we need to add the
`enableLegacyPatching` are done, and we can begin to plan the first
migration away from the legacy patching mode.

Bug: 1122511
Change-Id: I24e5dbb85e654eae28110ccfb603972034548921
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2517566
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>