Fixes dragging of the first column of a data grid by resizing
As part of adding undefined and null checks when migrating DataGrid.js to TypeScript (crrev.com/c/2482062), a check that incorrectly accounts for the number 0 as a falsy value was added to the resizer function.

This fix changes the check to compare with 'undefined' directly instead.

Fixed: 1145536
Change-Id: I09acb99eea77b389824562c1d3774b754755ba65
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2517157
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>