Check that block sections exists before add a blank style rule
Due to a race condition, the button to add a new style rule is
enable, but some data necessary to add the style rule may not
be ready yet.

I have tried to check the value of `_sectionBlocks`  to enable/disable
the add button, but even with that, the race condition can happen.

This will keep the same experience for users, but it will avoid
the exception.

Any idea will be more than welcome

Bug: 1142611
Change-Id: I5ead100cec3c20621c35212630511eab90c75de1
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2504403
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Jesus David Garcia Gomez <jegarc@microsoft.com>