Display description in console.table if preview is not available.
When calling console.table on native JavaScript objects, the _buildTableMessage method looks the use the 'valuePreview.properties' array as the values to pass into console.table.  Some native javascript objects present the 'valuePreview.properties' as an empty array causing no data to be passed or displayed.  To account for this, a second condition is added which preserves the original functionality of the method, and asks if the 'properties' array is empty, use the object 'default values'.

Bug: chromium:1057090

Change-Id: I99e97aeb79a184a1101003e1ad5d2a04dce63439
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2481736
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>