Add copy stacktrace contextmenu option for network initiator
If the initiator has a stracktrace, add the copy stacktrace
contextmenu option to automatically put the stacktrace on the
clipboard. Previously, users would have to select the text
in the tooltip, which could be problematic if the tooltip
disappears, because the mouse is outside the tooltip.

R=aerotwist@chromium.org

Fixed: 1139615
Change-Id: I9efebfc4909c0da5f8f0c2c46ebf7f191f3a6e75
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2523255
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>