Fix triggering search when clearing box with enter key
This bug was caused by the user tabbing to and hitting enter on the
"Clear" icon to clear the search result. Because the keydown event
listener was on the containing div, not on the actual search input, it
would fire a search event as if the user had hit enter from the input.

The fix is to move that event listener to being on the actual input
itself, which means we only re-fire a search when the user hits enter in
the input box.

Fixed: 1144615
Change-Id: I300a52e3a0af7e51ab28256bdd876881a865a38e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2523261
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>