Prevent auto-adding 'px' within CSS calc()
In a CSS calc() expression, it's enough that only one of the
numbers has a unit. E.g:

    calc(1px * 2)

Using the above example, in the previous behavior, incrementing
'2' by pressing the up arrow appends 'px' to it, resulting in
'3px'. This CL avoids adding the unit when incrementing/decrementing
a number in a CSS calc value.

Fixed: 1126236
Change-Id: Ief4d9941a8d90ed763f4dc225307e2c49365bd5c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2526022
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>