Settings: Allow users to create chord shortcuts
This CL adds the ability to create chords, or multi-keypress shortcuts,
to the shortcut editor. The ability for built-in chord shortcuts was
added here [1] in order to support those necessary for VS Code and
potentially other editor-matching presets (currently the only example is
Ctrl+K Ctrl+S to open the shortcuts settings with VS Code shortcuts
enabled).

The event listener that handles shortcut input keystrokes now starts a
timeout when it receives a non-modifier keypress, allowing the user to
begin a second key within 1000ms. If the user does, then the timeout is
cleared and the second key is added to the shortcut. Otherwise, any new
keystrokes after the timeout are considered the beginning of a new
shortcut entirely and the input's contents are replaced with the new
key.

[1] https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2125620
Video showing chord creation and timeout: https://i.imgur.com/wTjyxSg.mp4

Bug: 174309
Change-Id: Ic45fca8a5213f9befab6880c0c4ffbf41e22142a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2521142
Commit-Queue: Jack Lynch <jalyn@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Brian Cui <brcui@microsoft.com>