Display connection id '0' as empty cell in network log
The connection id '0' is used when there is no socket
associated with the request, or when it's not an HTTP
request. To make it more intuitive that 0 means essentially
"no information", this CL leaves the column blank if
"0" is the connection id.

Fixed: 1147811
Screenshot: https://imgur.com/a/h5zTxkg
Change-Id: I539f3353f29c50d7dd009da5622bfa4e18ce82f3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2532233
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Müller <petermueller@chromium.org>