Support expanding uppercased shorthand CSS properties
This bug was caused when the user passes in shorthand CSS properties
written in uppercase letters. The browser would not display the
'toggle-expandable' triangle indicating the availability of longhand
CSS properties.

The fix was to add toLowerCase() to the name argument inside of the
longhandsProperties() method. This allows the argument to be passed
into the longhands() method as a lowercase string, which then is able
to match the lowercase string and returns an array of longhand CSS
properties.

Screenshots:
- Before: https://i.imgur.com/qteeXcP.png
- After: https://i.imgur.com/Y9JzKQM.png

Fixed: chromium:918109
Change-Id: I018b3e7eccaa008ec2b9dbff101bb5681655e624
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2533305
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Mathias Bynens <mathias@chromium.org>