Add feature to LMI to operate on parts of a Uint8Array
Previously the whole Uint8Array needed to be passed down to the
linear memory inspector for rendering. This CL adds a
memoryOffset field to the Linear Memory Inspector to indicate
that the passed array is not complete and starts at the
given offset. If more data needs to be passed on a
navigation event (page, history), an event is triggered
to request the required data to be passed down.

Bug: chromium:1110202
Change-Id: Ide8cf2d85c1d31573d0c86a96968880e89a24ad8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2534192
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>