Fix <angle> keydown issue
This is caused by a recent change to preventDefault from mousedown,
which would have brought focus to the angleContainer. Now we have to
manually focus on it. Also we need to preventDefault on arrowUp/Down
to prevent scrolling.

Bug: chromium:1148183
Change-Id: Ic20858daebb374709ecf0bd011be720ca16c43b5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2534190
Commit-Queue: Changhao Han <changhaohan@chromium.org>
Auto-Submit: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>