Rename HighlightGridOverlay to PersistentOverlay
As the preparation for adding a persistent overlay for flexbox,
this CL renames HighlightGridOverlay to PersistentOverlay.
PersistentOverlay is expected to handle flex and grid overlays
in the future as both have to be shown together.

The same is done on the backend: https://crrev.com/c/2523175

Bug: 1145997
Change-Id: I929d9834084e99b212045db15c6a38fdc76f3378
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2523254
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>