[wasm] Turn DebuggerLanguagePluginManager into a singleton.
The `DebuggerLanguagePluginManager` used to be per `DebuggerModel` (and
hence per target) instead of being a singleton, actually a 1-1 companion
of the `DebuggerWorkspaceBinding`. This caused a number of problems,
starting with the fact that we'd never remove raw modules (the debugging
information held inside of the extensions) that were referenced by any
target other than the "main" target. It also came with a couple of hairy
inherent race conditions, and mandated that all extensions would perform
reference counting.

This CL addresses these problems by having only a single instance of the
`DebuggerLanguagePluginManager`, which listens for targets appearing
and disappearing, and thus centralizing all the communication with the
language plugins. This reduces both the memory overhead for bookkeeping
and also addresses the race conditions when talking to the language
plugins, since all of this goes through a central bottleneck now. Also
removes the cyclic dependency between extensions/ and bindings/ modules.

Fixed: chromium:1122000
Bug: chromium:1148763
Change-Id: I0eeb4ae88bfa618c35f73c8644187423aae3bea3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2537961
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Philip Pfaffe <pfaffe@chromium.org>
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>