Includes extra runtime look up into the new view extension infrastructure
To complete the basic infrastructure to migrate extensions of type 'view' away from module.json files, this refactors a look up for view extensions made to the runtime inside main/ to make it compatible with the new infrastructure for view extensions. This look up was missing in crrev.com/c/2508282.

Bug: 1134103
Change-Id: Id4f045bda4ef76637278b8ea86263b6426a9c254
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2519814
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>