[DataGrid]: enforce cell values are primitive
This CL enforces a stricter limit on what cell values are allowed to be.
I decided to allow `null` and `undefined` as that may well be something
that data grids may have - e.g. a cell might not have a value.

By enforcing the type we can also avoid enforcing a separate `title`
attribute and make that optional - if it's not provided we'll set the
`title` to be the stringified `value`. And we can then check in the sort
function for numbers and act accordingly, meaning if you provide numbers
the sort will just work, rather than having to provide a sort comparator
(which is a planned follow-up piece of work).

Bug: 1125968
Change-Id: I95d0fab51c836528b20e1c97a596721b83c598d8
Fixed: 1150329
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2546710
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alfonso Castaño <alcastano@google.com>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>