[Storage] Only delete first party cookies by default
This cl changes the default behaviour of removing all cookies that are
somehow related to the current page when clicking the "Clear site data"
button in the Storage tab of the Application panel. This lead to people
unexpectedly removing their session cookies on third party websites like
Google or Facebook when resources from them were included in the page.

To go back to the previous behaviour, this cl adds a checkbox next to
the button to also clear 3rd party cookies.

Screenshot: https://imgur.com/a/2PCz2gi
Bug:chromium:1012337
Change-Id: I171a6e3e07d56d5cbeafb0b2b030e3a5e24aad3b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2537954
Commit-Queue: Jan Scheffler <janscheffler@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>