Fix regression where color-picker isn't anchored on the color swatch
This bug comes from the fact that SwatchPopoverHelper expects an element
to anchor the color-picker to.
That element used to be the old color swatch component, which had a way
to get a reference to the icon element (the little colored square).

The new color swatch component doesn't expose its internal DOM structure
like the old one did, and it's probably better to keep it this way.
So the SwatchPopoverHelper is now getting the entire color swatch
component element as an anchor, causing it to position the color picker
at its center (so instead of pointing to the colored square, it points
to somewhere within the text).

In this change, I made it so the new color swatch component exposes a
public method that returns the AnchorBox instance that
SwatchPopoverHelper needs.

Before: https://imgur.com/eUJrvCk.png

After: https://imgur.com/SAk8dyO.png

Bug: 1152347
Change-Id: I1b79b6f4b50fdc1e3c6315d1e5fd239714cbf631
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2557841
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>