[New Linkifier] Add <devtools-linkifier> web component
This CL creates a new <devtools-linkifier> component that will eventually
replace the legacy linkifier code.

For now this component only supports taking a string URL, but in time (and as we
need) we can extend it to support all the inputs that the legacy Linkifier code
uses.

It works by emitting an event that is picked up by the legacy linkifier and
threaded through the legacy system.

Bug: 1149403
Change-Id: I76023052d818ab0f296db887fb03d6761ca4796b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2558297
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Alfonso Castaño <alcastano@google.com>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>