Remove custom setImmediate implementation
All of the existing callsites passed only a callback parameter, with
no additional `args` parameters, and so this functionality is
equivalent to the built-in queueMicrotask().

Note that there are some upstream tests that currently still use
setImmediate. These tests are ported in [1], which should be merged
before this patch can land.

[1]: https://chromium-review.googlesource.com/c/chromium/src/+/2563476

Bug: chromium:1050549
Change-Id: Iee3d286d38bcb17d9b78ca89f9897e8ca37b6d94
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2554560
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>