Prepare Trust Token report view for result section
This CL adds the trust token operation result to NetworkRequest and
forwards the data to the Trust Token tab in the Network panel.
As the operation result can arrive after the report was initially
rendered, this CL adds a new event to trigger a re-render as
necessary.

The CL doesn't yet add the rendering for the result itself, just
prepares the component to have the necessary data.

R=jacktfranklin@chromium.org

Bug: chromium:1126824
Change-Id: I9c6c080de9268219cd8b4b12804a7ca7c4de39a8
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2567783
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>