Flex: Remove temporary support for non-baseline flex overlay backend
Now that https://chromium-review.googlesource.com/c/chromium/src/+/2567643/
has landed in the backend, we can remove some code from the flex overlay
frontend.

The data the overlay backend was sending to the frontend has changed in
this CL, this is why I had introduced this temporary code that knows how
to deal with the the 2 data formats.
Now we don't need to deal with the old format anymore.

Bug: 1139949
Change-Id: I607eaa18a72a6a8143e3e9762709b3a781097a27
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2570745
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>