Add missing semicolons in TypeScript files
For all TypeScript files, we were still using the base semi-rule
from ESLint. However, that rule is unaware of TypeScript-specific
constructs. Instead, we should be using the typescript-eslint
version of the semi-rule (which extends the ESLint semi-rule), which
is unaware of these constructs.

This fixes the issue of missing semi-colons for type aliases and
in interfaces/field declarations. All pre-existing failures are
addressed in the same CL, since the odds of merge conflicts will
be low.

R=jacktfranklin@chromium.org

Fixed: 1154690
Change-Id: I986693a547dd17074bae3581660fa67b13a3b93a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2571118
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>