Fix flaky color swatch tests
Previously when ArrowDown was pressed it could happen that the body
tag was still collapsed. This lead to selecting the closing html
tag in the Elements panel instead of #inspected. This CL changes the
interaction to press ArrowRight two times that ensures that the body
tag is expanded and the #inspected element is selected.

Fixed: 1154590, 1149358
Change-Id: I0fbf4786713d115eb7018489d19cd78b1c0af7d3
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2573359
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>