[LMI] Add feature to jump to an address
On entering an address in the navigator the user can jump to the
memory at that address. This also fixes a few issues that showed
when activating this feature, e.g. what values to show in the
interpreter if we are at the last memory location.

Bug: chromium:1110202
Change-Id: I8c8b3ece103b6cec4b2f4ef31d704573898ae7a6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2571123
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>