Hides horizontal scrollbar in elements style panes
The horizontal scrollbar for the style sub-panes in the elements panel
originally had the horizontal scrollbar hidden as items wrap in case of
horizontal overflow. This change partially reverts crrev.com/c/2367354 which
had changed this behaviour, but we are still keeping the vertical scrollbar
fixed to avoid the offset to the right of the component.

R=tvanderlippe@chromium.org

Fixed: 1153267
Change-Id: Iaf3b8edb4193d04cbb68a6bf72e440e87b68459a
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2572674
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>