[LMI] Use arraybuffer-based types when opening the LMI
Before we would use the className property to determine whether or not
we are dealing with Uint8Arrays. This change now checks whether
we are dealing with arraybuffer-based types and allows the LMI
to open the view for any of these.

Also-by: bmeurer@chromium.org
Also-by: pfaffe@chromium.org
Bug: chromium:1110202
Change-Id: I12eeb8c5df846c96f93454b99ff32c6e88845e20
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2575089
Auto-Submit: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Philip Pfaffe <pfaffe@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>