Update frame tree when frames are written to via JS
Even though writing to a frame's document via JS is not considered a
navigation by the backend, it can change the information DevTools is
showing about a frame in the frame tree. We are adding a new CDP event
'Page.documentOpen' in the backend, which in the frontend triggers
the same cascade a frame navigation would. This ensures that
ResourceTreeFrame and therefore the frame tree is kept up to date.

Design Doc:
http://docs/document/d/1l-XzjFMdASjRzX4-vMxOoK-MXoxY5uhCe0GNfDsXeDo
Backend CL: https://crrev.com/c/2498482

DISABLE_THIRD_PARTY_CHECK=depedency

Bug: chromium:1140540, chromium:1140481
Change-Id: Ieb72c3ef0830aff19964310b1728d218d6da9943
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2519702
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>