Change extensions' `tags` field to array
Tags for an extension are declared in a single string and are
localized one by one at runtime. Tags to localize are provided to
the localization system by a static analysis done on module.json
files. However, with the new system for extensions, we cannot opt
for dynamically localizing tags because UI strings must be declared to
the localization system beforehand.

A workaround, implemented in this CL, is to provide an array of localized
strings as the tags of the extensions, where applicable.

Bug:1134103
Change-Id: I0760d47e25d346d4e52afad2ebdca8c014d63d86
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2577168
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>