[module.json extensions] Add support for ContextMenu providers
Modules can now register their contextMenu providers in the meta
files. The ContextMenu will now request each registered provider
alongside the previously registered providers from the runtime.

Note that a contextmenu provider will eagerly load its module, to
load the definition of the provider itself. Ideally we don't perform
this work at once (we currently load *all* modules that have a
contextmenu provider as soon as the user opens a contextmenu), but
we will punt that to a follow-up to clean up.

Lastly, some cleanups were performed in `show()` to use async-await
and handle the array of Promises that return an array of Providers.

R=andoli@chromium.org

Bug: 1134103
Change-Id: Idb15be2b6f4c9d8c6fb663f548580620500b076d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2569768
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>