[Accessibility]: Improve screen reader text to include count and position of suggest box (e.g. 'suggestion 3 of 5')
Issue:
- Suggest box does not announce count and position for screen reader users

Changes:
- Adding count and position to the SR text

Before: https://imgur.com/ZAXjEMX
After: https://imgur.com/Pfo3bi1

Why do we need to use ARIAUtils.alert to mimic screen reader behavior?

The issue here is that the suggestion box is not actually attached to the DOM of the text input like most listbox-type inputs.  SuggestBox uses a glassPane object which renders the suggestion box at the coordinates of the text input.  Because of this structure, focus is never actually placed on the suggestion box and therefore the screen reader cannot read the list properly.  We use ARIAUtils.alert to mimic what a screen reader would say given the context of the suggestion box and the selected item.

Bug: 1157329
Change-Id: Id668246ecc08b22e5b4ca86b1a767ba36b11c4f2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2582503
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Michael Liao <michael.liao@microsoft.com>