[surveys] Fix minor issues with survey links
Use an actual button instead of a link pretending to be a button. This
gives us Space + Enter functionality for the onClick handler for free,
and avoids the need for aria-role.

Change the focus styles so that we only get a focus outline when the
link is used with the keyboard, to avoid an outline on click.

Bug: 1112738
Change-Id: Ic9309ce95576c1fa8bd54220fdd8d1ca744bf804
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2584974
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>
Auto-Submit: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Simon Zünd <szuend@chromium.org>