Fixes unshown images larger than 1MB  in ImageView
As its name suggests, ImageView previews images in the inspected
site. By default, a data URL is used, but when the image is too big a
regular URL is used as the source of the previewed image instead. This
change adds an extra assigment to the 'src' property of the displayed
element in the case that the image is too large for using a data URL.
We lacked a proper handling for this case, which prevented any image at
all to be shown when its size exceeded a fixed length of 1MB.

R=aerotwist@chromium.org

Fixed: 1153229,1153611
Change-Id: I7bf3e50d494e0f4e3cdbcc7ff6ad3252c47f83ca
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2572661
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>