Use a ResourceTreeFrame's security origin as context in console
Previously, for generated OOPIF we selected the first available url
of the Runtime.CallFrame as a context in the console for context
selection. This change uses the security origin of the
ResourceTreeFrame instead.

We believe that the call frame url was only used because the
security origin could have been stale after programmatic updates
of frames via document.write. Since such updates are now accounted
for after the fix in https://crbug.com/1140481, we can rely
on the security origin to be correct.

The developer expectation for the subtitle is to see the origin of
the frame, and this CL is a step towards this goal.

Bug: chromium:1159332, chromium:1132260
Change-Id: I54d690075f6767f54da60c1dc203d00e6760b83f
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2593135
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>