[DataGrid]: test to ensure resize persists as data is added
I was wondering if this would work or not. Turns out it does (yay!) but
feels like the sort of thing that would be easy to break, so adding a
test for resizing, and then adding data, to ensure that adding data
doesn't reset the user's resize.

Change-Id: I0efb7def31119148e12e2c3decc726aaa8923425
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2593140
Auto-Submit: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>