Move accessibility-legacy to integration_test_runner
There are no further references to Accessibility in the module.json
files and therefore we are only reliant on it in the layout tests.
As such, we can remove it from the `modules` array and import
it in the integration_test_runner itself.

R=aerotwist@chromium.org

Bug: 1131500
Change-Id: Ia59741525a266659a9e37957057ba51d1674206e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2593142
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>