Fix Pretty-Print Infobar not working
This CL fixes the Pretty-Print Infobar not formatting the source file
when the Pretty-Print button is pressed. The bug was caused by a subtle
change to the button delegate when it was updated to support TypeScript
causing it to no longer invoke the format function.

The CL slightly modifies the ScriptFormatterEditorAction to have button
a event handler to absorb the unused 'event' argument, so the Infobar
does not need to pass one to toggleFormatScriptSource.

Bug: 1159531
Change-Id: I8e1876e8ce0ebc1e498e81a619a905a007cabdf5
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2596121
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Brian Cui <brcui@microsoft.com>