Emulation custom UA-Ch editor: fix some vertical alignment
The help button was missing the vertical-align property to make
it actually align with the text, so add that.

Also nearby fix alignment of the header line including the help
button and the folder triangle --- the triangle was aligned to contents
*including the margin*, so drop it, since things are spaced right
w/o it anyway.

Bug: 1160227
Change-Id: I474d43db62f05b41c922b79e42b1a7ae2a938de6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2597560
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Maksim Orlovich <morlovich@chromium.org>