Fix that cross-origin images cannot be saved
Currently, downloading images in imageView used `<a download>`, but Chrome 65[1] has begun to disable the download of cross-origin images.

Data URLs and Blob are not subject to this restriction, so this CL is changed to Data URL to save images. In the future, when the File System Access API matures, we can consider switching over.

[1]: https://www.chromestatus.com/feature/4969697975992320

Fixed: 1157399,1158212

Change-Id: I7e4929d0d531e029328ccd70ec91ef6d228b09cd
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2581801
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Changhao Han <changhaohan@chromium.org>