Change sources and profiler meta files imports to shell.js
'sources' and 'profiler' are delcared as modules of the 'shell' app
instead of 'devtools_app', thus the side effect imports of their meta
files must be located in shell.js not in devtools_app.js.

This CL corrects the location of these imports as their misplacement
prevents the appropiate loading of the dependencies of the 'shell' app.

Fixed: 1159888
Change-Id: I441149b657d86be4587c22843c7836be2f578e92
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2599754
Auto-Submit: Andres Olivares <andoli@chromium.org>
Commit-Queue: Yang Guo <yangguo@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>