Make sure rulers appear when needed
Rulers are supposed to appear around the viewport edges when the
showRulers setting is set.
This used to work but regressed at some point in the past. A lot of
refactoring occured on the overlay codebase, and some of it must have
broken this feature.

There is a check in the code to only draw the rulers once. This check is
called gridPainted (confusing because this has nothing to do with css
grid, but this was added 8 years ago in
https://chromium.googlesource.com/chromium/src.git/+/b2d7b736447d9d9d13466ad8ab07cfd4b27537d5)

This check made sense at the time to avoid redrawing on resize of the
page. But now, after refactoring, it no longer does this. Instead it
prevents the rulers from re-drawing at all after having been cleared.

It seems safe to just remove it entirely.

Bug: 1161701
Change-Id: Ibb67566ddcf699e3f6819267b48ec08f3f71408b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2601786
Reviewed-by: Brandon Goddard <brgoddar@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>