[i18n] De-duplicate identical i18n strings
These two are causing problems in the translations pipeline because they
are identical but have slightly different placeholders. The placeholders
don't convey a difference in semantics so we need to figure out if we
actually need separate strings for decimal places or not.

Right now this just removes the duplicate which doesn't change how the
translation would be done, it just makes the translation pipeline happy.

Bug: 1163928
Change-Id: Id640e9b07c315a11b339efd40b42066986f6d576
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2612863
Reviewed-by: Mathias Bynens <mathias@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>