Don't emulate screen size for fullsize shots to avoid unwanted resizing
We used to need to emulate a different screen size when taking fullsize
screenshots in devtools.

However, since https://bugs.chromium.org/p/chromium/issues/detail?id=1003629
was fixed, we now have access to a new CDP parameter when taking
screenhots that allows to create screenshots even when all or part of
the image is "below the fold".

This parameter, captureBeyondViewport, is already set to true in the
ScreenCaptureModel here:
https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/front_end/sdk/ScreenCaptureModel.js;l=55;drc=5215b2bdd2f7d5b1387df5783edbf9385591d221

This means the emulation data we're setting prior to taking the
screenshot is now redundant:
https://source.chromium.org/chromium/chromium/src/+/master:third_party/devtools-frontend/src/front_end/emulation/DeviceModeModel.js;l=834;drc=d7cfd148144b8bc0964ddfab1a2c91980519361c
and we can just remove it.

This fixes an issue where the resulting image was larger than the page
actually was.

Bug: 1158735
Change-Id: Ic44ea9ebea52165bdd231afb748dcc7b15d1d176
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2621732
Commit-Queue: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Robert Paveza <Rob.Paveza@microsoft.com>
Reviewed-by: Mathias Bynens <mathias@chromium.org>