[devtools] Remove timeline flow event experiment.
This experiment is relying on the deprecated format of flow events and
there are no plans to ship in the immediate future.

Remove it to unblock removing the old flow event macroses. This
experiment should be reimplemented on top of the new trace event flow
format instead.

R=yangguo@chromium.org,caseq@chromium.org
BUG=1164298

Change-Id: Ibf29a2842750c0fab8b6099e992aed27f4886b12
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2617780
Commit-Queue: Alexander Timin <altimin@chromium.org>
Reviewed-by: Yang Guo <yangguo@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>