Allow live locations even if the debugger is disabled
Previously we relied on the debuggerModel to create a raw location,
which is used for creating live locations.

This changes raw locations (DebuggerWorkspaceBinding.Location)
to rely on scriptIds instead of scripts. As a result we can create a
live location even if we currently do not have a script.

As soon as the debugger is enabled and a script is parsed, the
live locations will be updated with the relevant information.

We use this change in the Linkifier, which now shows no link if
the debugger is disabled (e.g. if we profile), and updates the
location and shows the link as soon as the debugger is enabled.

Also-by: bmeurer@chromium.org
Also-by: pfaffe@chromium.org
Bug: chromium:1132260

Change-Id: I48dc6016099d5b82d702a4331f93301693a6a6f9
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2595386
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>