[DevTools Frontend] Resync debugger id after navigation
DebuggerModel maintains a static table between debugger id and
DebuggerModel. This table is consulted by CallStackSidebarPane to
display call stack. If the mapping is not updated, call stack information
may not be displayed correctly, especially for async stack. There is a
need to resync the mapping, as a debugger id is mathematically based on
an isolate while a debugger model may work on different isolate.

Previously, we resync the mapping when we cannot find a debugger model
corresponding to an id. This change resync the id to model mapping when
responding to FrameNavigated event.

The work is based on conversation related to this CL:
https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2466642

Change-Id: I605e811b643ae154afb73024ea062f15b7872505
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2610136
Commit-Queue: Songtao Xia <soxia@microsoft.com>
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>