Fix: remove the underline from CSS var()
The underline was added to signal that the variable is a clickable link. It's visually too prominent though, especially with a lot of variables on the screen. I suggest to remove the underline completely, give the variable the `--link` color (blue) and show the underline only on hover. This is also acceptable in regards to a11y.

Screenshot: https://bugs.chromium.org/p/chromium/issues/detail?id=1147750#c11

Fixed: 1147750
Change-Id: Ic8a6bfcb78d4ed1b676fb6e114359701e9902f33
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2631126
Commit-Queue: Mathias Bynens <mathias@chromium.org>
Reviewed-by: Patrick Brosset <patrick.brosset@microsoft.com>
Reviewed-by: Changhao Han <changhaohan@chromium.org>
Reviewed-by: Mathias Bynens <mathias@chromium.org>