[startup] Wait until settings are loaded to load locale data
If we want to give users a setting to choose their locale, then we need
to wait until settings are available to load the locale data.

This CL reorders startup so that we load up settings/prefs first.
Also promisify prefs loading so that it's easier to follow ordering
in MainImpl.

Bug: 1167695
Change-Id: I0955e04a0b5f204c8784be2cad7e6b9fa5443b8b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2635657
Reviewed-by: Simon Zünd <szuend@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>