[module.json extensions] Adds getLazilyComputedLocalizedString to i18impl
Meta files defined for the new system for registering module extensions
are loaded before a locale is set for the i18n v2 system via the
i18n.registerLocale function. This in turn means that invocations to the
i18n system inside the meta files will throw errors as the locale is a
prerequisite.

As a workaround, a call to the getLocalizedString is wrapped in an
anonymous function so that the string is lazily localized when used.
However, this introduces a repetitive pattern that can be refactored into
a single function that when invoked returns the wrapped call. To avoid
repetition, such refactoring is introduced in this CL.

Bug: 1134103
Change-Id: I750b469002b6898143286d687088b90f71b2f00b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2638013
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>