Add network as dependency of security
The security panel can link to show requests in the network panel,
but did not specify it as a dependency in its module.json. This
caused CSS resources of the network panel to not be loaded and
cause a distorted view of the network panel.

R=andoli@chromium.org

Fixed: 1168071
Change-Id: Ifa9f6179d79ad2ffd0927db86db681765b476e21
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2642347
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>