[surveys] Make the survey prompt text injectable by the caller
As a drive-by, add a localized empty string constant so that
we can intialize class members with it, and so that we can avoid
having empty strings in multiple translation files as we do right
now.

Also rename the component_docs directory which still referred to
issues.

Bug: 1112738
Change-Id: I68bc11286e3b4c83f0a3bf4dbbcf1fb1932764b0
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2643508
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>