Show SameParty cookie attribute in cookies table
The SameParty attribute is a boolean cookie attribute used to indicate
whether a cookie should be included in requests to origins of the
same "first party set". This CL adds the attribute to the CookieTable
similar to how the "Secure" attribute is shown. Editing is not
yet supported.

Drive-by: Don't translate column headers that correspond 1:1 to
Set-Cookie attributes.

Screenshots: https://imgur.com/a/GoZj7UQ

R=sigurds@chromium.org

Bug: chromium:1161427
Change-Id: Ic208e8f1411fe1de755c42607190d187302ec21c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2623938
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>