Fix repaint bug for chrome remote debugging
This CL fixes a bug introduced in https://crrev.com/c/2445770. The
design of highlightInOverlay is misleading, as it should be called
updateHighlightInOverlayAndRepaint (because it is also responsible
for resetting the highlight, and, in the absence of a highlighted
node, for repainting the image). Another reminder that good method
names matter.

I'm deferring renaming the method to a later CL so we can keep this
small and back-mergeable.

Bug: chromium:1158340
Change-Id: Icd4b797e19fdb2d0136325a37d5855ef1f70cbcf
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2652510
Reviewed-by: Maksim Sadym <sadym@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>