Fix node action sources
When rewriting the actions that use Node to node_action,
we missed the "sources" to be forwarded to the action.
This means that a change to a source file does not properly
trigger a rebuild in GN, as GN doesn't understand it
is a source file.

R=jacktfranklin@chromium.org

Bug: None
Change-Id: I0313c6007a83417e1cc5566ae6d9bc379c17b337
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2653086
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>