[LMI] Use description from ArrayBuffer or WebAssembly.Memory as title.
With https://crrev.com/c/2653155 the inspector will expose the
WebAssembly.Memory for every ArrayBuffer as internal property (only for
ArrayBuffers and SharedArrayBuffers that are owned by WebAssembly
memories of course). This CL adds support for using that information in
the linear memory inspector to show the description of the
WebAssembly.Memory as title, falling back to the description of the
ArrayBuffer if it's a standalone ArrayBuffer.

Before: https://imgur.com/hod9jPR.png
After: https://imgur.com/v195VoC.png
Bug: chromium:1166577
Change-Id: Ibe2bde5a29588d53f50f25fb82f1254e60e44876
Fixed: chromium:1171621
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2652974
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>
Commit-Queue: Benedikt Meurer <bmeurer@chromium.org>