[WebAuthn] Fix focus when activating by spacebar
Before this patch, the checkbox got disabled to make sure that the
authenticator can not be disabled again while it is in the progress
of being enabled. This lead to the focus being lost.
With this patch, a new class member has been introduced to keep track
of the enabling instead.

Fixed: chromium:1167991
Change-Id: Ic5b0367d57a6001af6b77ce988eab207207149be
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2655249
Commit-Queue: Jan Scheffler <janscheffler@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>