Rename confusing function in ScreencastView
We had a function named _highlightInOverlay in ScreencastView,
the name of which makes it sound as if it was optional to call.
However, that function also handles repainting, which is not
optional, and this introduced the referenced bug.

This CL renames the function to make it somewhat clearer what
is going on.

Bug: chromium:1158340
Change-Id: I487828a92845426342be5063187ff3044f9bd682
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2656241
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>