[DataGrid] Ensure re-render if data updates mid-render.
In a previous CL we updated the data grid to render using the render
coordinator. As part of that work we also prevented multiple render
calls within the same frame. However, this resulted in a bug where data
would be updated in the same frame, and would be discarded.

This CL continues to block multiple renders being enqueued, but if the
data is updated (and render is called) during an existing render, we
schedule a render single render call at the end.

Bug: 1171202
Change-Id: Iccdb8770b1a93cc7810ae9bce18976e3c4bf92e2
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2659023
Commit-Queue: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Auto-Submit: Paul Lewis <aerotwist@chromium.org>
Reviewed-by: Jack Franklin <jacktfranklin@chromium.org>