[DataGrid] Don't attempt to re-focus column headers
We were being a bit keen when trying to re-focus a data-grid row if it
had lost focus due to being removed from the DOM and then added back in.
We don't need to apply that logic if the focused cell is a header,
because they are never removed from the DOM as the user scrolls.

Fixed: chromium:1173437
Change-Id: If27cbe70a4d3f7053581d0941869c30b827f2d57
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2667508
Reviewed-by: Simon Zünd <szuend@chromium.org>
Reviewed-by: Paul Lewis <aerotwist@chromium.org>
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>