Fix icon alignment in the Network Trust Token tab
A recent fix of the Icon component threw the success/failure icon in
the Trust Token tab out of alignment. This CL fixes that so that the
icon is baselined again with the first line of text (centering is not
straight-forward).

Screenshot: https://imgur.com/a/XceM0Xl

R=sigurds@chromium.org

Bug: chromium:1126824
Change-Id: I8e8f38ceb6faf77cb47829ffb75623dd6538b352
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2667509
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Commit-Queue: Simon Zünd <szuend@chromium.org>