Make emulation locations non-localizable and fixup incorrectly cached titles
The `emulation.locations` setting represents the locations visible in
the location settings tab. When the setting value is updated, it is
serialized and cached.

In crrev.com/c/2650505 we migrated the location emulation settings to
the emulation-meta.ts file, which included localizing all titles of
the locations. By doing so, the `title` field was updated from a normal
string to a function that returns a localized string. However, these
settings are serialized and cached and take the original values as input.
Since functions can't be serialized, the `title` would become empty in
the DevTools settings cache.

Since we don't need to have these titles localized at all, we can revert
back to strings rather than functions. This allows them to be stored
like normally in our cache. However, to fix any user that had these
location stored in the cache, we have to check if the title is empty
and patch up the location accordingly.

Fixed: 1171670, 1172220, 1172214
Change-Id: Ia403fbba1cca19b022cfb2b7d1b6f8dad768a889
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2666426
Commit-Queue: Andres Olivares <andoli@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>