[lmi] Reduce visual clutter on TypedArrays.
The previously added memory icon adds too much visual clutter to
TypedArrays which already come with an element preview. So this CL
restricts the memory icon to just ArrayBuffer and WebAssembly.Memory
for now, and leaves DataViews and TypedArrays unchanged. For the latter
it's easy to get to their buffers and click on the memory icon of the
buffer.

Bug: chromium:1166577, chromium:1174063
Change-Id: I316da8d079690f7917dfb0e4c71e69325a2f7f05
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2673822
Auto-Submit: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>
Reviewed-by: Kim-Anh Tran <kimanh@chromium.org>