Cleanup GRD input files
To facilitate the removal of `$resources_out_dir`, we should cleanup
the usages of that variable throughout the various GNI variables that
we define. Therefore, we perform the string concatenation of
`$resources_out_dir` in the GRD action, rather than in the variable
definition.

There are still numerous usages scattered throughout the codebase,
but we will need to tackle these separately, as that will involve
some more path wrangling.

R=alexrudenko@chromium.org,aerotwist@chromium.org

Bug: 1174013
Change-Id: I760c1b153f0d2e1f9cdf1098773b92a7d05259e6
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2678091
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>