[Linkifier] Fall back to url for raw location if no script exists
DebuggerModel.createRawLocationByScriptId will always succeed,
creating a raw location based on the script Id. The live location
that is created from that raw location will update all relevant
information required to open a link as soon as the script appears
in the debugger model.
However, some times a script is never resolved (no script parsed
event with that script id), and in this case we actually need to
find the correct script with a different scriptId using the url.

This adds a fall back in the Linkifier that considers to search for
a script using the url if the current raw location is not
linked to a script, but a script can be found via url.

Fixed: chromium:1173456
Change-Id: I5721bdc0549bb0516b07f285da32b3c688d2240d
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2678695
Reviewed-by: Benedikt Meurer <bmeurer@chromium.org>
Commit-Queue: Kim-Anh Tran <kimanh@chromium.org>