Fix width of right column in FrameDetailsView
The right column in this CSS grid has a `min-width`, which prevents the
'Owner Element' cell from growing if necessary. Instead of growing the
whole grid cell, it's just the button which grows and overflows. And as
a consequence the separators which are part of the grid, don't grow
along with the button.

This CL removes the cell's `min-width`, so that it grows along with the
button content.

Screenshots:
before: https://i.imgur.com/e2VsoHr.png
after: https://i.imgur.com/f8LMqB2.png

Fixed: 1175658
Change-Id: I28e6861e814029113b3a4cd5db91522dcb62143e
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2678697
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>