Fixing the order in the settings categories
Issue:
Items in settings categories appear in a different order from which they
were specified.

Root:
The conditional only returns an actual result of the comparisson if
the two elements in the category have an order, in any other case it will
threat them as equals.

Fix:
If the item has an order it will have higher precedence over items without
and order. e.g If a category has an order of 0 it will be placed at the
top.

Bug: 1176503
Change-Id: Ic7837c30aaa0a7a4f45a5bb7d9f14d6a51eb5f70
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2686810
Commit-Queue: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>
Reviewed-by: Andres Olivares <andoli@chromium.org>