[settings] Use new category enum for pre-ordered sections list
The SettingsScreen has a preference for the ordering of the setting
sections of the various modules. This list was previously hardcoded
with the english terms of the various categories.

After the module.json migration, the category names used the
translated names, causing a mismatch if a locale other then english
was used. For non-english locales, no settings would show up, as the
different sections did not exist.

This CL fixes this by using the previously introduced SettingCategory
enum to define the preferred ordering.

R=andoli@chromium.org, tvanderlippe@chromium.org

Bug: None
Change-Id: I18448b6cfe6338fb55e327644707b4999c99cd47
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2696328
Commit-Queue: Simon Zünd <szuend@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>