Fix flexbox editing for inline element styles
When inline element styles are being edited, it results in
DOM changes that cause all style sections to be rebuilt.
In this case, Flexbox Editor becomes attached to a style section
that is no longer active. Ideally, there should be a way to
fix the styles panel so that the editing lifecycle is predictable
but for now this CL rebinds the flexbox editor to the new
style section when such an rebuild happens.

Bug: chromium:1175699
Change-Id: Ic49fdfcb6e71d27ca2177dcb5173c4e787330d5c
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2698840
Commit-Queue: Alex Rudenko <alexrudenko@chromium.org>
Reviewed-by: Michael Liao <michael.liao@microsoft.com>