Allow ReportView content to be selected (and copied)
This makes the text content of ReportView selectable, so that it can be
copied and pasted elsewhere.

Not only the values themselves, but labels and headers are selectable
too. This allows users to easily copy and search for terms they might
not be familiar with.

Bug: 1093247
Change-Id: I33cf9f82c4f209dcca969953b2662f6e02dcdc18
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2698842
Reviewed-by: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Peter Müller <petermueller@chromium.org>
Commit-Queue: Wolfgang Beyer <wolfi@chromium.org>