Fix i18n computation on module scope
These strings are computed in the module scope, which could mean that
the localization system hasn't finished loaded yet. Instead, they
should be lazily computed, so that when the strings are used, the
localization system is available.

This change is required to unblock loading `timeline` in `MainImpl`
for an extension that we are porting.

R=petermarshall@chromium.org

Bug: none
Change-Id: I1b3573d1d5c3154b74920d410a372ee566e0321b
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2699266
Auto-Submit: Tim van der Lippe <tvanderlippe@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Peter Marshall <petermarshall@chromium.org>