[settings] Show memory settings if there are any visible
The memory setting 'memoryLiveHeapProfile' has no associated UI and
thus can't be enabled. It is only visible/usable if the associated
experiment LIVE_HEAP_PROFILE is enabled, too.

This CL adds a memory section to the preferences page, and also adds
code to filter out empty sections in the preferences page so we don't
have a heading with no settings under it when the above experiment is
turned off, which looks awkward.

Bug: none
Change-Id: Iab7882a7ffb4612405bd85af18c4751cff04c626
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2703954
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>