[TreeOutline] tell custom renderers if the node is expanded
We need this for the object viewer because when we show objects we show
different data if a node is expanded.

For example, if we are looking at an object with a property that's not
expanded, we might show:

```
{
  foo: {...}
}
```

But when we expand it we show:

```
{
  foo:
    * bar: 2
}
```

So we need to dynamically render the `{...}` part only if the node is
not expanded.

Bug: chromium:1166669
Change-Id: I45ca851b1d6bdfb11c012a573ec0656cf1bcf644
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2703959
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>