Show allowed/disallowed features in FrameDetailsView
This is a premiminary surface for the new CDP permissions policy
queries, and it doesn't show the reason a feature is disallowed yet,
but we think it will be useful anyway. For the next release, we plan
to report the blocked reasons including references to the relevant
headers/iframe attributes.

Bug: chromium:1158827
Change-Id: Ie4ef47914952e229706447b1bd7edebba13eac95
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2698841
Commit-Queue: Sigurd Schneider <sigurds@chromium.org>
Reviewed-by: Wolfgang Beyer <wolfi@chromium.org>