[l10n] Update and clean up description strings for elements/
Doing a manual pass over all the descriptions in elements/ to make sure
they are good enough.

Also refactor the SplitWidget so that it does not concatenate localized
strings. Previously it would substitute in the name of the particular
sidebar panel that it was showing/hiding. This would have led to poor
translations in some languages.

Refactor it to accept the show and hide UI strings as input, and change
all the callers to provide both. Also update the types to
LocalizedString so it is type safe now re. localization too.

Bug: 1163928
Change-Id: Ie67db90770ff76ab258f885e1fe39387a26ab887
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2705378
Reviewed-by: Alex Rudenko <alexrudenko@chromium.org>
Commit-Queue: Peter Marshall <petermarshall@chromium.org>