[TreeOutline] take a generic `key` data type.
The TreeOutline implementation previously took nodes with primarily one
piece of data to represent the current node, and an array of children to
denote a node's child elements:

```
{
  key: string,
  children: [...],
}
```

This works fine for basic TreeOutlines where you're just rendering
string nodes, but as I discovered when trying to put together an Object
Viewer this structure is fairly rigid and does not allow for more
complex trees.

Let's imagine we try to render a tree for a JSON object that contains:

```
{
  name: 'jack',
}
```

Converting this into a tree node, you might get:

```
{
  key: 'name:"jack"',
}
```

But you don't want to render it as a plain string; we want to style the
property and the value differently to make it clearer to the user what
we're representing. At this point you might make use of the custom
rendering functionality to do this:

```
{
  key: 'name:"jack"',
  renderer: node => {
    const [property, value] = node.key.split(':');
    return LitHtml.html`...`
  }
}
```

This works, but has limitations in that we're forced to take this object
and represent it as a string, only to then parse its parts out of that
string representation later on. I then experimented with allowing Nodes
to take optionally some additional `data`, of any type. This means we'd
represent `{ name: 'jack' }` as:

```
{
  key: 'name:"jack"',
  data: {
    property: 'name',
    value: 'jack'
  },
  renderer: node => {
    const {property, value} = node.data;
    return LitHtml.html`...`
  }
}
```

This works, and is much better. We can store the actual data we want and
use it without a temporary string representation in the middle. But what
do you notice about this Node? `key` is entirely unused!

This lead me to the realisation that `key: string` was far too
simplistic; and rather than allowing people to attach arbitrary data to
a Node (which is also challenging to keep type-safe), I should instead
make the Node's `key` property entirely generic (note that this is a
slimmed down example, not the actual code):

```
interface TreeNode<DataType> {
  key: DataType;
  renderer: (node: TreeNode<DataType>) => LitHtml.TemplateResult;
  children?: TreeNode<DataType>;
}
```

Doing this still allows for the very basic trees (`TreeNode<string>`)
but allows us to create more complex trees too and keep type safety.
What I like about this approach is demonstrated in the
`tree_outline/custom-renderers` component docs example, just with this
data change I can build out a functioning CSS properties viewer just by
providing data and a custom renderer: https://imgur.com/a/Ab0kFv4.

The one downside is _a lot_ of types now need to become generic; but I
think this is a very worthwhile trade-off.

Bug: chromium:1166669
Change-Id: I5e0d4b407c5e9fc2a414722a7a1b225b27440088
Reviewed-on: https://chromium-review.googlesource.com/c/devtools/devtools-frontend/+/2705085
Commit-Queue: Jack Franklin <jacktfranklin@chromium.org>
Reviewed-by: Tim van der Lippe <tvanderlippe@chromium.org>